Fix bug with FetchHashKeyName usage propagating to SHOW CREATE TABLE statement
A change in FetchHashKeyName within get_trigger_data would eventually impact
the mixed-case 'Create Table' column hash key later on in get_table_sql.  The
attr is a dbh-only one, so this seems best solved with a localized set in any
spot where we're using one of the select*_hashref methods.